+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+* Handler profiling is now disabled by default.
+
 ## 0.93.18
 
 * [Feature] Add an option to profile VPN handler execution: if enabled, a warning will be written to the log